Assignment 2-Interactive Comic-Boredom to Adventure
BOREDOM TO ADVENTURE: A CHOOSE-YOUR-OWN JOURNEY THROUGH TIME

Project Description
"BOREDOM TO ADVENTURE: A CHOOSE-YOUR-OWN JOURNEY THROUGH TIME" is an interactive "choose your own story" comic, coded and designed using HTML, CSS, and JavaScript.

This comic is the story of a kid who was unhappy with his life and embarks on a trip through different eras and futures as he goes from the dinosaur era to the Jazz Age of the 1920s, to an ultramodern futuristic city to a dystopian nuclear future - ultimately leading the character to realize that his everyday life is not so bad as he initially thought.

The central theme revolves around embracing one's present life and finding joy and excitement within it. The comic seeks to engage users not just as passive recipients of a story, but as active participants shaping their narrative. The experience is designed to be immersive with the audios that were used and the user interactivity.
Process
The development of the project was a step-by-step process that started with brainstorming the storyline.

Storyboarding: We all sat down together and brainstormed and wrote down the storyline and a vague sketch of each panel.

Design: Then Sunny and I collected the images from the web and she edited them on Adobe photoshop

Sound: I collected all the sounds and edited them on Audacity.

Coding: Then we broke down the code and each took a part to make the website using HTML, CSS and JavaScript to structure, style and make the pages interactive.

Reflection/Evaluation
Reflecting on the project, I feel that it was a success and I was able to do exactly what I had in mind. The interactive comic creates an engaging environment that lets users decide their path, while subtly conveying the central message.

Building this project was a fantastic learning experience. Not only did I improve my skills by doing new things such as adding audio to code and using Audacity, but I also delved deeper into the interplay of narrative, interactivity, and user experience. Of course there were challenges that at first seemed impossible to fix but thanks to the group we were able to achieve the majority of what we had in mind in the beginning minus some stuff that no matter what we did just wasn’t working out like for example adding sound to the buttons (we tried several different attempts and methods but none seemed to work and so we decided to move on feeling discouraged since after exactly 4 hours of trial and error we just couldn’t do it). But in the end, the comic came out amazing and in my opinion reflects the “comic principles” by having a beginning, middle and end, by having a conflict, etc…